# spec/generated/openapi.yaml
openapi: 3.0.0
info:
  title: Calendar Slots API
  version: 0.1.0
  description: 'Запись на звонок: владелец публикует типы встреч, гость бронирует свободный слот.'
tags:
  - name: EventTypes
  - name: Bookings
paths:
  /api/bookings:
    get:
      operationId: listBookings
      description: Предстоящие встречи всех типов, ближайшие первыми.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
      tags:
        - Bookings
    post:
      operationId: createBooking
      description: Гость занимает свободный слот; занятое время даёт 409 даже для другого типа встречи.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
  /api/event-types:
    get:
      operationId: listEventTypes
      description: Каталог типов встреч для гостя и владельца.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
      tags:
        - EventTypes
    post:
      operationId: createEventType
      description: Владелец создаёт тип встречи.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - EventTypes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventType'
  /api/event-types/{id}:
    get:
      operationId: getEventType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - EventTypes
  /api/event-types/{id}/slots:
    get:
      operationId: listSlots
      description: Свободные слоты на 14 дней вперёд, начиная с текущей даты.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - EventTypes
components:
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - not_found
            - validation_error
            - slot_unavailable
            - already_exists
        message:
          type: string
      description: Ошибка с машинным кодом и сообщением для пользователя.
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - eventTypeTitle
        - start
        - end
        - guestName
        - guestEmail
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        eventTypeTitle:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        guestName:
          type: string
        guestEmail:
          type: string
        createdAt:
          type: string
          format: date-time
      description: Встреча в календаре владельца.
    BookingRequest:
      type: object
      required:
        - eventTypeId
        - start
        - guestName
        - guestEmail
      properties:
        eventTypeId:
          type: string
        start:
          type: string
          format: date-time
          description: Начало выбранного слота из списка свободных.
        guestName:
          type: string
          minLength: 1
          maxLength: 120
        guestEmail:
          type: string
          pattern: ^[^@\s]+@[^@\s]+\.[^@\s]+$
      description: Данные гостя для записи на слот.
    EventType:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          type: string
          maxLength: 64
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          description: 'Идентификатор-слаг, задаёт владелец: `intro-call`.'
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        durationMinutes:
          type: integer
          format: int32
          minimum: 30
          maximum: 240
          description: 'Длительность встречи: кратна 30 минутам, от 30 до 240.'
      description: Тип встречи, который владелец календаря предлагает гостям.
    Slot:
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      description: Свободный интервал для записи. Время в UTC.